+++ packages/media-parser/src/containers/iso-base-media/collect-sample-positions-from-moof-boxes.ts
@@ -1,0 +1,26 @@
+import {getSamplesFromMoof} from '../../samples-from-moof';
+import type {IsoBaseMediaBox} from './base-media-box';
+import type {TfraBox} from './mfra/tfra';
+import type {TkhdBox} from './tkhd';
+
+export const collectSamplePositionsFromMoofBoxes = ({
+	moofBoxes,
+	tfraBoxes,
+	tkhdBox,
+}: {
+	moofBoxes: IsoBaseMediaBox[];
+	tfraBoxes: TfraBox[];
+	tkhdBox: TkhdBox;
+}) => {
+	const isComplete =
+		tfraBoxes.length > 0 &&
+		tfraBoxes.every((t) => t.entries.length === moofBoxes.length);
+
+	const samplePositions_ = moofBoxes
+		.map((m) => {
+			return getSamplesFromMoof({moofBox: m, trackId: tkhdBox.trackId});
+		})
+		.flat(1);
+
+	return {samplePositions: samplePositions_, isComplete};
+};
