+++ packages/media-parser/src/containers/iso-base-media/get-sample-positions-from-track.ts
@@ -1,21 +1,10 @@
-import {isLpcmAudioCodec} from '../../get-audio-codec';
-import {getTimescaleAndDuration} from '../../get-fps';
 import type {SamplePosition} from '../../get-sample-positions';
-import {getSamplePositions} from '../../get-sample-positions';
-import {getSamplePositionsFromLpcm} from '../../get-sample-positions-from-lpcm';
-import {getSamplesFromMoof} from '../../samples-from-moof';
 import type {IsoBaseMediaBox} from './base-media-box';
+import {collectSamplePositionsFromMoofBoxes} from './collect-sample-positions-from-moof-boxes';
+import {collectSamplePositionsFromTrak} from './collect-sample-positions-from-trak';
 import type {TfraBox} from './mfra/tfra';
 import type {TrakBox} from './trak/trak';
-import {
-	getCttsBox,
-	getStcoBox,
-	getStscBox,
-	getStssBox,
-	getStszBox,
-	getSttsBox,
-	getTkhdBox,
-} from './traversal';
+import {getTkhdBox} from './traversal';
 
 export const getSamplePositionsFromTrack = ({
 	trakBox,
@@ -32,67 +21,20 @@ export const getSamplePositionsFromTrack = ({
 	}
 
 	if (moofBoxes.length > 0) {
-		const isComplete =
-			tfraBoxes.length > 0 &&
-			tfraBoxes.every((t) => t.entries.length === moofBoxes.length);
+		const {isComplete, samplePositions} = collectSamplePositionsFromMoofBoxes({
+			moofBoxes,
+			tfraBoxes,
+			tkhdBox,
+		});
 
-		const samplePositions_ = moofBoxes
-			.map((m) => {
-				return getSamplesFromMoof({moofBox: m, trackId: tkhdBox.trackId});
-			})
-			.flat(1);
-
-		return {samplePositions: samplePositions_, isComplete};
-	}
-
-	const isLpcm = isLpcmAudioCodec(trakBox);
-	const timescaleAndDuration = getTimescaleAndDuration(trakBox);
-
-	if (isLpcm) {
 		return {
-			samplePositions: getSamplePositionsFromLpcm(trakBox),
-			isComplete: true,
+			samplePositions,
+			isComplete,
 		};
 	}
 
-	const stszBox = getStszBox(trakBox);
-	const stcoBox = getStcoBox(trakBox);
-	const stscBox = getStscBox(trakBox);
-	const stssBox = getStssBox(trakBox);
-	const sttsBox = getSttsBox(trakBox);
-	const cttsBox = getCttsBox(trakBox);
-
-	if (!stszBox) {
-		throw new Error('Expected stsz box in trak box');
-	}
-
-	if (!stcoBox) {
-		throw new Error('Expected stco box in trak box');
-	}
-
-	if (!stscBox) {
-		throw new Error('Expected stsc box in trak box');
-	}
-
-	if (!sttsBox) {
-		throw new Error('Expected stts box in trak box');
-	}
-
-	if (!timescaleAndDuration) {
-		throw new Error('Expected timescale and duration in trak box');
-	}
-
-	const samplePositions = getSamplePositions({
-		stcoBox,
-		stscBox,
-		stszBox,
-		stssBox,
-		sttsBox,
-		cttsBox,
-	});
-	if (samplePositions.length === 0) {
-		return {samplePositions: [], isComplete: false};
-	}
-
-	return {samplePositions, isComplete: true};
+	return {
+		samplePositions: collectSamplePositionsFromTrak(trakBox),
+		isComplete: true,
+	};
 };
